[js-api] Fix `WebAssembly.Memory` buffer accessor
A given agent will only have one `WebAssembly.Memory` object that
references a given `Data Block`. Even if a `Memory` is `postMessage`'d
to another worker and back, the agent's memory object cache will contain
a reference to the `Memory` object associated with that `Data Block`.

So if another agent shares a memory with this agent, and the other agent
grows the memory, the only thing that's different in this agent is the
`Data Block`'s length. So we check that in the getter, and if it's
different we create a new buffer object and store it on the cached
`Memory` object.