[interpreter] Add `initThreads` hook to JS output
A threaded test requires a more substantial harness than can be provided
in one JavaScript file, and must be implemented differently for Web JS
engines vs. console-based JS engines.

The idea is that the surrounding harness will implement the
initThreads() function, and replace the `instance` and `register`
functions to send messages to the Worker threads.

The `initThreads()` function is async so it can block until the threads
are ready. The entire function is wrapped in an `async` function to
allow for the `join` command (not yet implemented) to block.