Add a note for WebAssembly.Memory.prototype.grow return value (#158)
The current text specifies the return value to be the current size of memory in pages (before resizing), but this is somewhat ambiguous and can lead to implementation specific behavior. For example, engines can decide to return the value at the time the grow method was called, or at the time the memory was updated, and these can be different values depending on the strictness of the implementation itself, and if there are multiple grow calls that occur at the same time. 

Add a note to clarify that the return value should essentially be an atomic RMW to the length of the underlying buffer. The implication
of this is that m ore than one grow call cannot return the same value.