Fix error in Web embedding desc for atomic.notify (#185)
In https://github.com/WebAssembly/threads/pull/110 we changed the `count` argument of `atomic.notify` so that it is treated as an unsigned `i32`. Previous to this change, it was interpreted as a signed `i32`, with negative numbers signalling that all threads should be woken. This part of the overview wasn't fully updated - fixing this now.