Fix some typos in `*.wast` tests
Most of the test's `$Check` modules imported from `"mem"` which was
failing when I integrated this into Wasmtime. This may very well be a
bug in my understanding of the `*.wast` format, however, but currently
via `(module $Mem ...)` that defines the name `"Mem"` for modules to use
but not `"mem"`. If this is supposed to work as-is I can update Wasmtime
instead.

Additionally some tests' comments were checking `L_1` but were actually
checking against `local.get 0` so I've updated them to `local.get 1`
which fixed some failing tests on Wasmtime.